requirements.txt
seldom==2.3.3
requests
jmespath
jsonschema
attr
pyrsistent
six
selenium
urllib3
colorama
chardet
certifi
idna
xmlrunner
jinja2
markupsafe
parameterized
openpyxl
jdcal
pymysql
jsonpath
ruamel.yaml